remote_services/requirements.txt
fastapi
pillow
pymupdf==1.25.5
loguru
ngrok
uvicorn
python-dotenv
python-multipart
unsloth
vllm
google-genai
firebase-admin